Update patch set 3
Patch Set 3:

(1 comment)

nit: if the rc file is not landed on the target, we should detect that and report back to the developer what they may be missing for testing pre-requisites (assuming test run on older devices than this CL, or newer devices should we decide to gate the landing/existence of the file on some other factors). Will consider if a respin is required (two dependent patches in series).

Patch-set: 3
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>