Set system first_api_level from vendor image.
Shipping API level is associated with the vendor image. When
a device ships in, say, O and is then run against the O-MR1
GSI image, the shipping API level is currently reported as
the shipping API level which the GSI was built with. This means
that various tests and control flows assume that the image
is the newest possible image. However, this is not the case.

Setting the system property here until it can be officially
deprecated.

Bug: 67965044
Test: boot on pixel
Change-Id: I2f66705a8afca4fa3035156703ea7b3b2a1e958f