Update patch set 2
Patch Set 2:

(1 comment)

I understand the bug and your fix, but please find a way to do this with flags_, or on a cleaned up state interface.  

Adding a secondary status_ variable is going to take an already error prone situation (flags_) and compound it.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>