Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> First of all, I agree on you that this change is correct. However, with this change, linker restrictions on VNDK-SP are _not_ enforced on all devices which don't set BOARD_VNDK_VERSION. 

Correct, since the VNDK-SP set does not exist unless BOARD_VNDK_VERSION was set.

> We have been enforcing the restriction on VNDK-SP since O and that has been part of the Treble compliance testing. So, I don't want to agree on this change.

The makefile creating the fake vndk-sp libraries was approved with the condition that it was for prototyping only, and would never reach shipping android devices. Since the only way that this worked without BOARD_VNDK_VERSION was with that code, we shouldn't have been doing this.

> If our immediate goal is to make aosp_walleye bootable, we can simply copy libdrm.so into vendor partition (as suggested in the bug), or change libdrm back to vendor:true so that it is installed to vendor partition even when BOARD_VNDK_VERSION isn't set.

That's not possible -- there is not a /vendor/libdrm.so without BOARD_VNDK_VERSION.

Patch-set: 1
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>