Update patch set 1
Patch Set 1:

> That's not possible -- there is not a /vendor/libdrm.so without BOARD_VNDK_VERSION.

This is because libdrm is vendor_available:true, right? I think we can change it back to vendor:true and define another libdrm (i.e. libdrm_platform) for libminui.

Anyway, aosp_walleye will be bootable again when O-MR1 is published to AOSP and thus aosp/master is rebased. (since it can then be built with BOARD_VNDK_VERSION). So, I am okay with this change if this will be reverted back then.

I do agree that the fake vndk-sp was nasty, but I think we have to keep it until we stop supporting Oreo version of vendor image.

Patch-set: 1
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>