Update patch set 1
Patch Set 1: Code-Review-2

I really don't want to introduce a separate flag to describe these two cases.  Collectively they're trying to express to the OS that a storage location is "stable" and that signal is used for both making a device adoptable and for making it visible to apps, and there should never be a difference between the two.  If we introduce two flags, there's a pretty high risk that OEMs will misconfigure their devices by setting one and forgetting the other.

Patch-set: 1
Reviewer: Gerrit User 1013508 <1013508@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2