Update patch set 1
Patch Set 1:

Do you have an example of where the right context is failing to be set?  Shouldn't the right context be applied right when the symlink is made in the normal case?

This might be the case in /dev or before restorecon has happened to the parent directory, so I'll be curious to see the example.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>