Update patch set 2
Patch Set 2:

> No, not yet triggered, it was caught as a critical issue by static code scanning. 
> 
> if that is the case, how to gracefully catch this exception (page fault)? do you have best practice? thanks

Android endeavors to keep a certain amount of memory available for system processes, killing apps as needed to maintain it.  So it really shouldn't happen.

If you want to make sure, though, you could pre-allocate a buffer of the largest possible size and reuse it.  Keymaster is single-threaded so there's never any need for more than one message buffer and one response buffer.

Patch-set: 2