Update patch set 1
Patch Set 1:

(18 comments)

Thanks for the contribution, this is one of Android's ToDo items (but we did not write up a backing-bug for it, only hallway discussions). The efficiency gains in testing did not place it at a high priority because there are more efficient alternatives, but given that you are running a reader for each bucket, makes sense to cover an inefficient alternative.

We _strongly_ recommend that you deploy _one_ reader logcat -b all -D and use the supplied dividers to drop the contents into individual buckets inside your development collection mechanism. Using that approach reduces the CPU utilization even more than with this adjustment (other vendors have used this approach). With your change citedhere, you might be tempted to continue as-is, please do not. We would prefer limiting the number of individual processes and logger connections.

Please look into logpersist daemon, which can be switched on and off in the Developer Options. logpersist is strictly controlled and only delivered on userdebug systems and has been covered by a security review. logpersist does -D and lands the content into /data/misc/logd/ which can be collected (in logpersist selinux context or in a bugreport). It offers a 256MB buffer horizon, but does have minor impact storage wear (unknown how much).

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>