Update patch set 1
Patch Set 1:

(2 comments)

Steven, as we talked, we need native binaries in odm partition as it is the customization of the vendor partition. For example, there might be some HAL services in the odm that replace the ones in the vendor partition.

Patch-set: 1
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>