Update patch set 1
Patch Set 1: Code-Review-1

I don't think this is worth it.

I'd like to keep this as library as small as possible given that the zip format is a popular target for security exploits. I agree this seems safe but if we're not sure, it's best to do nothing.

Patch-set: 1
Reviewer: Gerrit User 1014443 <1014443@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1