Update patch set 2
Patch Set 2:

do we run the risk here of not noticing next time the kernel is broken and just falling back to the slow method forever?

or is the "sometimes" in the checkin comment not referring to kernel changes but some specific subset of processes (because of minijail or whatever)?

Patch-set: 2
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>