Introduce property schemas
Properties right now can take any format, but that makes it hard to
specify an API for these properties as Treble intends to do.
Therefore this change introduces the idea of property schemas.  There
are 4 possible schemas defined below:

1) 'any' this is the default schema and allows any property to be set.
2) 'bool' this allows only boolean values (true|false|1|0)
3) 'int' and 'uint' these allow signed and unsigned integer values
respectively.  These optionally take a set of parameters to specify a
min and max value that can be set, e.g. 'int -3 3' or 'uint 0 1'.
4) 'enum' this allows only a specific set of space deliminated values
to be set, e.g. 'enum allow these strings' only allows one of 'allow',
'these', or 'strings' to be set.

Test: unit tests
Change-Id: I7a6b00fb43ec630d1f56c9e9a1f1b61d3914f603