Update patch set 1
Patch Set 1: Code-Review+2

I think this is a reasonable work around for now. I am working on moving this code to the new unwinder, and it understand the offset extended opcode so it shouldn't need to have to do this.

Patch-set: 1
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2