Update patch set 1
Patch Set 1:

This is the wrong change for this issue.  The place where this clock is being used should be changed to using std::steady_clock which is the C++ wrapper around CLOCK_MONOTONIC instead.  This class is specifically meant to provide access to CLOCK_BOOTTIME.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>