Update patch set 1
Patch Set 1: Code-Review+1

(2 comments)

nits: The expected(?) test failures land in CTS and go/apct. We need those that see the failure to get a clear message that the test failed _because_ of a high rate background selinux spam that needs to be addressed. I (as the most notable) do not want to be drawn into a bug that really should be addressed by the platform.

Even if the test is _not_ flaky as a result of background activities (now that it no longer has the dynamic ratelimiter hysteresis, maybe it also works "better"), we should report a public message to whomever reads it, that the background rate is too high?

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1