Statically link lmkd to reduce mlockall bloat
We also define a new liblog subset to avoid pulling into lmkd random
C++ goo. We set a stack rlimit to avoid the problem, of some systems,
the kernel allocating a huge stack all at once (8MB) and locking it
all into memory. (The stack should be allocated incrementally, but in
practice, this incremental allocation doesn't happen on some systems.)

Change-Id: I6cc40a2c23b010d21817c61d15a16a715a177b35