Update patch set 2
Patch Set 2:

How much of these savings are r/w pages vs text segment?  I'm generally a fan of preferring C++ to C in all cases, so I'm a bit skeptical of this change, especially whether it'll be future proof if you need to link to other libraries in the future.  And especially if some of these savings are text segment pages that would otherwise be shared with other processes on the system.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>