Update patch set 3
Patch Set 3:

Not sure if it's required in AOSP CLs but since you have a bug for these features would be good to have "Bug: 71605857" in the descriptions of these CLs. That I think will also update the bug automatically when CL is merged.

Patch-set: 3
Reviewer: Gerrit User 1157738 <1157738@85c56323-6fa9-3386-8a01-6480fb634889>