Update patch set 2
Patch Set 2:

(6 comments)

Split this CL, too much might be going on here? Here are my suggestions based on the (non-existent) architecture document :-) :

The first one defers lock+thread initialization until the first enable() is called. Before that point all one has is the populated operations and no resources taken. The assumption is that only one process is making the calls to keep this simple; but the deferring of the initialization is a first step towards preventing a force suspend handler from bringing up multiple threads?

The second one (or maybe this is the third one if it is not expected to happen) uses a global exclusive flock() (?) to test and prevent multiple instances of autosuspend thread running? The risks will always be _which_ thread is active, so an enable()/disable() in the other process may have to also have a inter-process channel to communicate this, so another exclusive flock() may be expended to block/enable autosuspend. This must be accompanied by support for inter-process communications I guess. This _complication_ is not worth the fury if your expected usage pattern is to place this all into Power Manager. 

The real meat with regards to splitting things up is the start with a KISS force suspend, then grow it based on technical need:

The third one introduces force suspend with timeout and wakesource exclusion mechanism (?), but a KISS version that ignores the timeout, takes a local wakelock, clears the non-excluded wakesources, and goes straight to "mem" > /sys/power/state. You will also need to introduce a means to register which wake sources must be excluded, which may be an API or HAL to libsuspend, or some property that libsuspend can read. The possibilities are endless, and the options are an architectural decision for you to make considering all levels of the services involved. One possibility (but complicates jni, or Power Manager implementation?) could be a callback function passed to force suspend (with a default one if it is nullptr that uses common power key matches) that is passed the wake source string, and returns a true/false as to whether to exclude it or not. Maybe it is supplied an exclusion list as an array and size to perform the matching directly? The caller is expected to have quiesced all wakelocks (eg: Display is blanked, but for now do add the Hail Mary set display brightness to zero?) and turned off all non-essential peripherals (eg: GPS, sensor service/sensor hub) to minimize power.

NB: We could move the protocol of managing the wake sources higher up than libsuspend, if it is accompanied by a means to disable autosuspend before (powering down and) disabling the wake sources.

The fourth one introduces support for a timeout. Not strictly required for day one, but I bet(tm) that some partner will grumble about the rug being pulled out from under the feet of all the processes should they be asynchronously performing user space cleanup, or have pending timeouts that may trigger in a bad sense in a flurry on return. Only perform the lock+thread initialization if the timeout supplied is non-zero. The timeout is there to make sure a piece of locked up hardware or HAL is non-blocking, so as to not block the expected eventual lower power state. An error reported back is up to the caller, but the expected response would be to power-down if suspend to RAM is not possible.

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>