liblog: test: add liblog.time_drift
Confirm that MONOTONIC_RAW has no drift per se, and that MONOTONIC,
REALTIME and BOOTTIME have very similar direction of drift.  Sample
for 20 seconds, and at each end grab 32 time values for each to
average them out to compensate for variance in durations to collect
the time information.  We print the calculated drift information in
estimated seconds per year (0.032 ppms).

The accuracy of the information collected was expected to be within
10% only, after performance compensation.  Accuracy based on
experimentation on a device (marlin) without a fixed CPU clock rate
and varying conditions to achieve that goal with a 50% engineering
margin to help ensure the test is not flaky.  If a more accurate
result is required, you can lock the cores or select performance.

Test: liblog-unit-tests --gtest_filter=liblog.time_drift
Bug: 69423514
Change-Id: I2715dd439d2cfa05f83ff130746a6da64696c9a6