Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Why do you want to put these benchmarks here?  Probably better in a kernel benchmark suite that would be run when kernel changes are made?

Yup, want them there too, VTS tests, utilizing the linux tests, do cover bits and pieces of this. Getting them upstream will be more problematic and even further from liblog. Bionic benchmarks are too far away too. They needed to be localized for the continuing device-per-device analysis that I have hopes of landing in go/apct.

[TL;DR]

These are directly related to the bug investigation, and directly impact the performance of liblog (and logd). These are not necessarily kernel tests, they are in fact integration benchmarks (so kernel + bionic, but as they pertain to liblog time accuracy) localized for developer convenience and focus.

Will be adding a liblog test (and in go/apct) that runs the benchmarks (somewhat guaranteed to exist because of same directory when one does mmma system/core/liblog) and reports if the vdso is not operating correctly. The addition of these tests is exceedingly small in the context of liblog and can be run every time a developer has to with any changes. Connecting tests to benchmarks is still problematic in the context of go/apct. Using a corresponding kernel or bionic to complete may be an even bigger exercise in futility ...

These are also in bionic/benchmarks (except for CLOCK_MONOTONIC_RAW), running them is more problematic from liblog tests and we ran into strong headwinds to add them. We still have a few of the CLs pending that are not getting any traction. Adding more without the corresponding automated analysis and fault thresholds (proposed in liblog) got me back to here. These _are_ the historical origins of the bionic benchmarks I added, I did find them silly at the time here and went on a two month long quest to get them into bionic benchmarks. I miss the convenience, building bionic benchmarks to investigate this bug started making me feel silly that they were _not_ also here. Additional benchmark or tests are being explored to measure comparative drift between the clocks, which is the root of the bug issues with kernel and user space correlation accuracy.

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0