Update patch set 5
Patch Set 5:

Anything that uses libunwindstack as a static library will need to add libdexfile as well. Looks like some of the debuggerd libraries may be the offenders in this case, though there are a few other cases in the tree too.

Patch-set: 5
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>