libcutils: test: Add AshmemTest.scan_direct
If the ashmem protection afforded to us by ashmem_valid is being
bypassed by a direct ioctl call to ashmem, lets confirm that the
error return from the ioctl agrees with ashmem_valid.

This test is more of a kernel security check, failures will point
to issues that should be inspected in drivers _other_ than ashmem.

Test: libcutils_test --gtest_filter=AshmemTest.ashmem_direct
Bug: 72021458
Change-Id: I04bf97cad4041d4d92d01d216f07edc8d71d81c5