Update patch set 4
Patch Set 4: Presubmit-Verified+1

=== Finished presubmit run: L03100000137509225 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/595068?revision=4&workplanId=L03100000137509225




TreeHugger finished with: 23 passed (10 new).
















WARNING: The tip-of-tree(4563063) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4562379), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.










WARNING: The tip-of-tree(4563063) for git_master/taimen-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4562349), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4563063) for git_master/walleye-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4562349), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





I am TreeHugger, and I have hugged this change!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 4
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=+1