Update patch set 5
Patch Set 5:

There is a healthy margin on all the benchmarking thresholds, we are really looking for inversion (syscall takes less time than the vdso call), I ran this test overnight repetitively (unfortunately did not have a counter of how many times it ran) and the tests had zero failures on a Hikey960. I will need to run this on a pixel device and check if the tests ever flake out (this time, shell script to add a count of recorded successes and failures) over at least a 24 hour period.

Patch-set: 5
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0