Update patch set 1
Patch Set 1:

(5 comments)

Some nits we could cover given that lmkd is supposed to be fast, accurate, mlocked and miserly on memory usage. If we are going to have to deal with merge conflicts, might as well be in for several fixes?

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0