Update patch set 1
Patch Set 1:

The alluded-to CL that I'm aware of was never merged anywhere.  If you have a reference to the Android N CL you based this on, please post the Gerrit URL or git SHA.

New charger names appearing after boot isn't a common feature -- if some platforms need it then we can make the update call available from the healthd HAL, but we might not want to call it on every battery/charger update for every platform.

Patch-set: 1
Reviewer: Gerrit User 1004277 <1004277@85c56323-6fa9-3386-8a01-6480fb634889>