Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

I'd prefer for this change to be property-based and opt-in. We're trying to stamp out all instances of userdebug behaving differently from user, and I think atrace from a path like this counts as undesired difference between user and userdebug.

Also, please mention "ftrace" or "atrace" or something in the commit message. "Trace" is an overloaded word and it's not immediately clear which one is meant here.

Patch-set: 1
Reviewer: Gerrit User 1130394 <1130394@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1