WIP: Cache JIT debug entries.
This should allow debug-malloc to cache JIT debug entries
between unwinds and to look up needed entry quickly by PC.

This is important for debuggable apps as JIT is used extensively for
them. On the other hand, it is of little significance for production
apps, since those don't generate any JIT debug info at the moment.

Test: 137-cfi
Change-Id: I87ba70322053a01b3d5be1fdf6310e1dc21bb084