Update patch set 1
Patch Set 1:

This is not to allow settable accessibility to vendor components, but to allow partners to use the property as an action trigger in vendor init scripts.

So partners can't change the value of "vold.decrypt" with this CL.

The request was received through an email, so I wasn't able to add the bug id. I will file one and specify their rationale.

Patch-set: 1
Reviewer: Gerrit User 1019198 <1019198@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0