init: usb: Remove previously created symlinks
When switching between ConfigFS USB compositions the
on sys.usb.config=none is triggered but the current actions
only remove the f1, f2 and f3 symlinks. As a result if
the previously enabled composition supported more than 3
functions (which is possible with config_oemUsbModeOverride)
the additional symlinks will remain present. This will cause
the next activation to include those stale functions or
even fail to enumerate as ffs.adb could be included twice.

Fix this issue by always removing the symlinks for f1..f5. This
of course assumes there would be at most 5 functions enabled.

Change-Id: I4ad6a9d718c15fd325717834a665437e80042e1b