Update patch set 6
Patch Set 6: Code-Review+2

This is another change that we need to empirically verify. I'm skeptical of the special case for go devices. At the very least, we should condition behaviors on characteristics of the memory workload and available resources. With respect to memory behavior, "go device" is a circumstance, not an environment.

Patch-set: 6
Reviewer: Gerrit User 1130394 <1130394@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2