Update patch set 3
Patch Set 3:

yeah, it doesn't matter in practice. I'm fine with merging this as is. There was just another case where someone was loading and unloading every HIDL interface implementation in a directory (long since reverted). We fixed the HIDL libraries so they are hardened against this and work, but assuming that it's done in any percentage of libraries (in Android or elsewhere) would be a *big* assumption. Anyway, I think I realized what I can do to make it better.

Patch-set: 3
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0