Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 1:
> > 
> > "When DbC is initialized and enabled, it will present a debug device through the debug port (normally the first USB3 super-speed port)."
> > 
> > Does the debug port work if the Android is connected to a superspeed hub? Which port would it choose as the debug port?
> 
> We can add hub to the Debug host. A more detail information of the topology is in xhci specification 1.1 in section 7.6.1.

Our desired use case is to add a hub to the debug target. I see in the spec that this is currently specifically disallowed.

"A Debug Target may only connect to a Debug Host through a Root Hub port of the target."

Is this a technical limitation of a usb hub? One of our desired use cases is to debug a Android in host mode while it is connected to a device (such as usb headphones). Since pretty much all android phones have only one usb port, a hub is needed.

Patch-set: 2
Reviewer: Gerrit User 1113594 <1113594@85c56323-6fa9-3386-8a01-6480fb634889>