Update patch set 1
Patch Set 1:

This isn't wrong per se, but it's not exactly right either.  Didn't we add the ability for vendors to supply their own configuration for fs_config as opposed to modifying this one for files on the vendor partition?

Init has the ability to specify capabilities when services are launched too, which is probably easier than dealing with setting the capability flags in the file system.

Patch-set: 1
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>