Update patch set 1
Patch Set 1:

Ah, I was just searching for the string itself.

Is 'usbradio' a common boot mode or something specific for these devices?  I'm guessing the latter, which still makes me feel like this is mostly a 'vendor' property.

Are the devices in question going to be using gadget hal?  I assume so.  Can we prefix this with vendor for them?

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>