Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Is there not a better place for this file too though?  Why not put
>  > all of this in bionic/linker/etc or something closer to the
>  > consumer itself?  I'm not a huge fan of having a generic source
>  > folder for what goes to /etc, especially when these files have a
>  > specific consumer, whose source they can be moved close to.
> 
> ah, i see what you mean. yeah, that's not a bad idea. the only other file in here is "/etc/hosts" which is arguably part of bionic too.
> 
> anyone have any suggestions for where the linker configuration files should live?

yeah.. I agree. this is not the ideal place; far from the consumer and also these are actually not in rootdir(/) but under /system.

bionic/linker/etc sounds good, and I would like to suggest bionic/linker/configs and device/generic/common. Rationale for the latter; the linker config itself is not part of bionic though its syntax is. The content of the configs are dependent on some other subsystems like RenderScript, so I personally think it isn't ideal to have the configs in a specific module but in a neutral place.

public.libraries.[android|wear].txt doesn't seem to be part of bionic; it is read by libnativeloader. /system/core/libnativeloader/ (with no subdir) would be fine.

Patch-set: 1
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>