adb: support deduplicated ext4 in adb remount
When ext4 partitions are deduplicated, the kernel will refuse to mount
them as writable (since it does not know how to un-deduplicate shared
blocks). We handle this situation in adb remount by forcing an e2fsck on
the partition, since e2fsck understands how to undo the deduplication.
The user is asked to reboot the device before trying adb remount again.

Bug: 64109868
Test: adb remount works with and without ext4 deduplication
Signed-off-by: David Anderson <dvander@google.com>
Change-Id: I2ee9855704bcfc351c2ff3417fb2985b36509f89