Update patch set 1
Patch Set 1:

What does this help with?  The bug reference healthd having an issue logging to dmsg, but the comment here says that init is the only user of KernelLogger?

Is there an associated change that passes /dev/kmsg to any services to populate the environment variable?

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>