Update patch set 1
Patch Set 1:

This daemon does not need to run as root, and once it is not root, does not need the capabilities service command. Experimented with moving lmkd to a new user "lmkd", the capabilities appears to be working. Passes the current unit test.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>