Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> is this widely used already? it's a terrible API that keeps growing extra random arguments, and still doesn't cover a lot of the would-be use cases.
> 
> there are plans for a new API in Q if the point of this is to offer new functionality rather than to keep old code building...

Enabling vendor_available just makes it easier for partners to use the lib. Otherwise, they will need to fork one in their own source tree (as they are required not to change the AOSP projects). A new API in Q is certainly welcome. This CL will be the best interim solution before that happens.

Patch-set: 1
CC: Gerrit User 1061057 <1061057@85c56323-6fa9-3386-8a01-6480fb634889>