Allow specifying a boot time offset
This change adds support for the ro.boot.boottime_offset property, which
specifies the point in time (in the CLOCK_BOOTTIME clock) in which the
device actually booted. This is useful for containers, which are unable
to reset their clocks so that their concept of "uptime" matches when the
init process started executing.

Bug: 77273909
Test: CtsBootStatsTestCases
Change-Id: I9559aca8d9a7a36e40c691aa393c3df040a758d4