Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > (2 comments)
> > 
> > Shouldn't the reboot be automatic? Can system stability be assured while the fsck is running and after?
> 
> I was worried that an automatic reboot on remount would be unexpected. If you think it's not a big deal I'll add it. I do think it would be safer.

I was worried about that myself. One option would be to do something like this:

Have adb remount just _fail_ by default on a deduped filesystem. This way, there's no danger and no unexpected reboot. Then, separately, add some kind of "-R" flag to adb remount that means, dedup or not, "reboot immediately after performing the remount", and then let adb remount -R complete successfully on a deduped FS.

You'd implement -R by setting a persistent system property, then rebooting. Early in startup, before anything really runs, we'd check the system property, and if it's set, remount read-write. If we can't do that because the filesystem is deduplicated, we run fsck at _that_ point (when it's safe, since nothing else is running) and reboot again, automatically.

This way, there's no window of system instability, and behavior becomes uniform and predictable whether or not the FS is deduplicated.

It might be a lot of work to do it this way though.

Patch-set: 3