Update patch set 1
Patch Set 1:

This is what a tool would look like that could start a service as init.  In writing it though, I'm realizing that this is not a good idea.  There are so many little things that init does that are hard to replicate and I can see this becoming bit-rotted very easily.  

https://android-review.googlesource.com/c/platform/system/core/+/666381 is probably the real solution here.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>