[RFC] libsync: Add sw_sync_ functions for drm_hwcomposer
The drm_hwcomposer implementation uses libsync and some of the
sw_sync_* functions, which causes trouble when building with
Treble.

I'm not sure what the best approach is here, but this patch
adds the needed sw_sync functions to the libsync.map.txt
which allows the linker to find them.

Suggestions for better approach would be welcome!

Change-Id: I5ac29ebc240a7149620666b2615b4b6386e8659c
Signed-off-by: John Stultz <john.stultz@linaro.org>