Prepare to fail in RefBase destructor if count is untouched
Move towards crashing if a normally configured RefBase object is
destroyed without ever incrementing the reference count. We've been
threatening to do this for a long time. The previously last known
violation had been fixed.

This also fixes stack trace printing from RefBase, which had previously
been broken, and which we found necessary to track down further
violations of this rule. Unfortunately, we found several more violations
with the aid of that fix. After existing CLs are submitted, there are
still some failures, but they are no longer numerous. This this CL
doesn't actually crash in the event of a violation, but does log a
verbose stack trace if it encounters one.

The plan is to ensure that bugs will have been filed against the
remaining known offenders before submitting this CL, and we will then
enable crashing on usage violations once those are fixed.

The fix for the stack trace printing breakage unfortunately requires
the use of weak symbols in order to avoid a circular build dependency.

Some random reformatting, driven by consistency with current formatting
requirements.

Add missing include to BacktraceMap.h.

Bug: 79112958
Bug: 30292291
Test: Boot AOSP, Master
Change-Id: I8151c54560c3b6f75ffc4c48229f0388a2066958